+++ src/main/java/com/devcourse/ReviewRanger/user/application/CustomUserDetailsService.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Component;
 
 import com.devcourse.ReviewRanger.common.exception.RangerException;
+import com.devcourse.ReviewRanger.user.domain.UserPrincipal;
 import com.devcourse.ReviewRanger.user.repository.UserRepository;
 
 @Component
@@ -22,6 +23,7 @@ public class CustomUserDetailsService implements UserDetailsService {
 	@Override
 	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
 		return userRepository.findByEmail(email)
+			.map(UserPrincipal::new)
 			.orElseThrow(() -> new RangerException(FAIL_USER_LOGIN));
 	}
 }
