+++ src/main/java/com/devcourse/ReviewRanger/user/domain/User.java
@@ -2,9 +2,6 @@ package com.devcourse.ReviewRanger.user.domain;
 
 import static com.devcourse.ReviewRanger.common.regex.UserRegex.*;
 
-import java.util.Collection;
-import java.util.Collections;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -13,19 +10,14 @@ import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
 
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
 import com.devcourse.ReviewRanger.BaseEntity;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import lombok.Getter;
 
 @Getter
 @Entity
 @Table(name = "users")
-public class User extends BaseEntity implements UserDetails {
+public class User extends BaseEntity {
 
 	public enum Role {
 		ROLE_USER,
@@ -60,40 +52,8 @@ public class User extends BaseEntity implements UserDetails {
 		this.role = Role.ROLE_USER;
 	}
 
-	@JsonIgnore
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		GrantedAuthority authority = new SimpleGrantedAuthority(this.role.name());
-		return Collections.singleton(authority);
-	}
-
-	@JsonIgnore
-	@Override
-	public String getUsername() {
-		return this.name;
-	}
-
-	@JsonIgnore
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@JsonIgnore
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@JsonIgnore
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@JsonIgnore
-	@Override
-	public boolean isEnabled() {
-		return true;
+	public void updateInfo(String editName, String editEncodedPassword) {
+		this.name = editName;
+		this.password = editEncodedPassword;
 	}
 }
