+++ src/main/java/com/devcourse/ReviewRanger/user/domain/UserPrincipal.java
@@ -1,0 +1,66 @@
+package com.devcourse.ReviewRanger.user.domain;
+
+import java.util.Collection;
+import java.util.Collections;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
+public class UserPrincipal implements UserDetails {
+
+	private final User user;
+
+	public UserPrincipal(User user) {
+		this.user = user;
+	}
+
+	@JsonIgnore
+	@Override
+	public Collection<? extends GrantedAuthority> getAuthorities() {
+		GrantedAuthority authority = new SimpleGrantedAuthority(user.getRole().name());
+		return Collections.singleton(authority);
+	}
+
+	@JsonIgnore
+	@Override
+	public String getUsername() {
+		return user.getEmail();
+	}
+
+	@JsonIgnore
+	@Override
+	public String getPassword() {
+		return user.getPassword();
+	}
+
+	@JsonIgnore
+	@Override
+	public boolean isAccountNonExpired() {
+		return true;
+	}
+
+	@JsonIgnore
+	@Override
+	public boolean isAccountNonLocked() {
+		return true;
+	}
+
+	@JsonIgnore
+	@Override
+	public boolean isCredentialsNonExpired() {
+		return true;
+	}
+
+	@JsonIgnore
+	@Override
+	public boolean isEnabled() {
+		return true;
+	}
+
+	public Long getId() {
+		return user.getId();
+	}
+}
